requirements.txt
asgiref==3.5.2
Django==4.0.6
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
environ==1.0
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
